Added data augmentation
